smartpantry/backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
email-validator
python-multipart
google-genai
pytesseract
Pillow
